backend/requirements.txt
annotated-types==0.6.0
anyio==3.7.1
click==8.1.7
ecos==2.0.12
et-xmlfile==1.1.0
exceptiongroup==1.1.3
fastapi==0.104.1
h11==0.14.0
httptools==0.6.1
idna==3.4
joblib==1.3.2
numexpr==2.8.7
numpy==1.24.2
openpyxl==3.1.2
osqp==0.6.3
pandas==2.1.2
pip==22.2.2
pydantic==2.4.2
pydantic_core==2.10.1
python-dateutil==2.8.2
python-dotenv==1.0.0
python-multipart==0.0.6
pytz==2023.3.post1
PyYAML==6.0.1
qdldl==0.1.7.post0
scikit-learn==1.2.2
scikit-survival==0.21.0
scipy==1.10.1
setuptools==63.2.0
six==1.16.0
sniffio==1.3.0
starlette==0.27.0
threadpoolctl==3.2.0
typing_extensions==4.8.0
tzdata==2023.3
uvicorn==0.23.2
uvloop==0.19.0; sys_platform != 'win32'
watchfiles==0.21.0
websockets==12.0
